pyproject.toml
[project]
name = "ai-proposal-generator"
version = "0.0.1"
description = "A proposal generator pipeline with LLM agents."
authors = [
    { name = "Anonymous", email = "noreply@example.com" },
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "django>=5.0",
    "python-dotenv>=1.0.0",
    "drf-yasg>=1.21.8",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "social-auth-app-django>=5.4.2",
    "drf-spectacular==0.28.0",
    "djangorestframework-simplejwt>=5.4.0",
    "pillow>=11.1.0",
    "django-cors-headers>=4.6.0",
    "gunicorn>=20.1.0",
    "psycopg2-binary>=2.9.10",
    "boto3>=1.36.16",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langsmith>=0.1.0",
    "requests>=2.31.0",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "apscheduler>=3.10.4",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.7",
    "python-magic>=0.4.27",
    "langgraph[all]==1.0.0",
    "langgraph-cli[inmem]>=0.4.7,<0.5.0",
    "python-toon>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0,<5.0.0",
    "tomli>=2.2.1,<3.0.0",
    "black>=25.1.0,<26.0.0",
    "isort>=6.0.1,<7.0.0",
    "flake8>=7.3.0,<8.0.0",
    "mypy>=1.18.1,<2.0.0",
    "bandit>=1.8.6,<2.0.0",
    "pydocstyle>=6.3.0,<7.0.0",
    "codespell>=2.4.1,<3.0.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["src"]}}
[tool.setuptools.package-dir]
"agents" = "src/agents"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_errors = true
